audit-service/requirements.txt
# FastAPI & ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# Cryptography
cryptography==42.0.0
PyNaCl==1.5.0

# HTTP & Utils
python-multipart==0.0.6
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0

# Optional: Search
meilisearch==0.28.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Linting
black==23.12.1
ruff==0.1.9
